# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервис для платформы копитрейдинга",
        "title": "Copy Trading API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/strategies": {
            "get": {
                "description": "Возвращает список стратегий с пагинацией и фильтрами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Список стратегий",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Минимальный ROI",
                        "name": "min_roi",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Максимальная просадка",
                        "name": "max_drawdown_pct",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Оценка риска",
                        "name": "risk_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Количество записей на странице",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/strategy.StrategyListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новую торговую стратегию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Создать стратегию",
                "parameters": [
                    {
                        "description": "Данные стратегии",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/strategy.CreateStrategyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/strategy.Strategy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/strategies/{id}": {
            "get": {
                "description": "Возвращает стратегию по её идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Получить стратегию по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/strategy.GetStrategyByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные стратегии по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Обновить стратегию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/strategy.UpdateStrategyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/strategy.Strategy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/strategies/{id}/status": {
            "patch": {
                "description": "Изменяет статус стратегии (active, archived, deleted)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Изменить статус стратегии",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/strategy.ChangeStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/strategy.Strategy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/strategies/{id}/summary": {
            "get": {
                "description": "Возвращает суммарную статистику по стратегии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Получить сводку по стратегии",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/strategy.StrategySummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.StrategyStatus": {
            "type": "string",
            "enum": [
                "preparing",
                "active",
                "archived",
                "deleted"
            ],
            "x-enum-varnames": [
                "StrategyStatusPreparing",
                "StrategyStatusActive",
                "StrategyStatusArchived",
                "StrategyStatusDeleted"
            ]
        },
...
        "strategy.ChangeStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "enum": [
                        "active",
                        "archived",
                        "deleted"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.StrategyStatus"
                        }
                    ]
                },
                "status_reason": {
                    "type": "string"
                }
            }
        },
        "strategy.CreateStrategyRequest": {
            "type": "object",
            "required": [
                "account_id",
                "nickname",
                "user_id"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "avatar_url": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "payment_account_id": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "strategy.GetStrategyByIDResponse": {
            "type": "object",
            "properties": {
                "active_subscriptions": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/common.StrategyStatus"
                },
                "title": {
                    "type": "string"
                },
                "total_commissions": {
                    "type": "number"
                },
                "total_copied_trades": {
                    "type": "integer"
                },
                "total_profit": {
                    "type": "number"
                },
                "total_subscriptions": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "strategy.Strategy": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "master_account_id": {
                    "type": "integer"
                },
                "master_user_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/common.StrategyStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "strategy.StrategyListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/strategy.Strategy"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "strategy.StrategySummary": {
            "type": "object",
            "properties": {
                "strategy_id": {
                    "type": "integer"
                },
                "total_profit": {
                    "type": "number"
                }
            }
        },
        "strategy.UpdateStrategyRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "payment_account_id": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}